requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Markdown==0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.60.1
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.49.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
SQLAlchemy==2.0.44
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
